feat: centralized patch system
