+++ cairo/tests/ethereum/cancun/test_fork_types.py
@@ -4,16 +4,14 @@ from ethereum.cancun.fork_types import EMPTY_ACCOUNT, Account
 from ethereum_types.numeric import U256, Uint
 from hypothesis import example, given
 
-from mpt.ethereum_tries import EMPTY_BYTES_HASH, EMPTY_TRIE_HASH
-from tests.utils.strategies import account_strategy
+from keth_types.types import EMPTY_BYTES_HASH, EMPTY_TRIE_HASH
 
 
 class TestForkTypes:
     def test_account_default(self, cairo_run):
         assert EMPTY_ACCOUNT == cairo_run("EMPTY_ACCOUNT")
 
-    # Not using ellipsis, as hypothesis loads the strategies before loading the EELS patches.
-    @given(account_a=account_strategy, account_b=account_strategy)
+    @given(account_a=..., account_b=...)
     @example(
         account_a=Account(
             nonce=Uint(1),
