+++ python/cairo-addons/src/cairo_addons/testing/hooks.py
@@ -42,7 +42,8 @@ logging.basicConfig(
 )
 logger = logging.getLogger()
 
-pytest_plugins = ["cairo_addons.testing.fixtures"]
+# Require the `keth_types_patches` plugin to apply patches before anything else
+pytest_plugins = ["keth_types_patches", "cairo_addons.testing.fixtures"]
 
 
 def parse_int(value):
